♻️ Refacto Add components to be more maintainable
