+++ components/Employee/Event/table/Header.vue
@@ -1,0 +1,36 @@
+<template>
+<tr>
+  <th
+    scope="col"
+    class="py-3.5 pl-2 inline-flex group truncate text-left text-sm font-semibold text-gray-900 sm:pl-4"
+  >
+    Id
+  </th>
+  <th
+    scope="col"
+    class="py-3.5 pl-2 text-left text-sm font-semibold text-gray-900 sm:pl-4"
+  >
+    <span
+      class="inline-flex truncate group"
+    >
+      Nom de l'événement
+    </span>
+  </th>
+  <th
+    scope="col"
+    class="px-3 py-3.5 text-left text-sm font-semibold text-gray-900"
+  >
+    <span
+      class="inline-flex truncate group"
+    >
+      Statut
+    </span>
+  </th>
+  <th
+    scope="col"
+    class="relative py-3.5 pl-3 pr-4 sm:pr-6"
+  >
+    <span class="sr-only">Actions</span>
+  </th>
+</tr>
+</template>
